pyproject.toml
[project]
name = "movifisio"
version = "0.1.0"
description = ""
authors = [
    {name = "Joao Pedro Schiavoni",email = "joaopschiavoni@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.136.0,<0.137.0)",
    "uvicorn (>=0.45.0,<0.46.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "python-multipart (>=0.0.26,<0.0.27)",
    "alembic (>=1.18.4,<2.0.0)",
    "sqlalchemy-utils (>=0.42.1,<0.43.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "pydantic[email] (>=2.13.3,<3.0.0)"
]

[tool.poetry]
packages = [{include = "movifisio", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.3,<10.0.0)"
]
